Вынес некоторые функции из routes в отдельный файл
