+++ src/node/api/__init__.py
@@ -1,2 +1,2 @@
 
-from . import routes
+from . import routes, routes_debug
