+++ src/node/node_server.py
@@ -6,12 +6,14 @@ from threading import Thread
 from node.settings import constants
 import pymongo
 from flask import Flask
-from node.api.routes import bp
+from node.api.routes import bp as routes
+from node.api.routes_debug import bp as routes_debug
 from pymodm.connection import connect
 
 def start_server(port):
     app = Flask(__name__)
-    app.register_blueprint(bp)
+    app.register_blueprint(routes_debug)
+    app.register_blueprint(routes)
     app.run(port=port)
 
 if __name__ == "__main__":
